feat(widgets): remove useless conversions

+++ SharpEngine.Core/Widget/Button.cs
@@ -123,29 +123,29 @@ public class Button(
 
         if (_state == ButtonState.Hover && Active)
             SERender.DrawRectangle(
-                (int)(RealPosition.X - (Size.X + 4) / 2),
-                (int)(RealPosition.Y - (Size.Y + 4) / 2),
-                (int)(Size.X + 4),
-                (int)(Size.Y + 4),
+                RealPosition.X - (Size.X + 4) / 2,
+                RealPosition.Y - (Size.Y + 4) / 2,
+                Size.X + 4,
+                Size.Y + 4,
                 Color.White,
                 InstructionSource.UI,
                 ZLayer
             );
 
         SERender.DrawRectangle(
-            (int)(RealPosition.X - Size.X / 2),
-            (int)(RealPosition.Y - Size.Y / 2),
-            (int)Size.X,
-            (int)Size.Y,
+            RealPosition.X - Size.X / 2,
+            RealPosition.Y - Size.Y / 2,
+            Size.X,
+            Size.Y,
             Color.Black,
             InstructionSource.UI,
             ZLayer + 0.00001f
         );
         SERender.DrawRectangle(
-            (int)(RealPosition.X - (Size.X - 4) / 2),
-            (int)(RealPosition.Y - (Size.Y - 4) / 2),
-            (int)(Size.X - 4),
-            (int)(Size.Y - 4),
+            RealPosition.X - (Size.X - 4) / 2,
+            RealPosition.Y - (Size.Y - 4) / 2,
+            Size.X - 4,
+            Size.Y - 4,
             BackgroundColor,
             InstructionSource.UI,
             ZLayer + 0.00002f
@@ -168,10 +168,10 @@ public class Button(
 
         if (_state == ButtonState.Down || !Active)
             SERender.DrawRectangle(
-                (int)(RealPosition.X - Size.X / 2),
-                (int)(RealPosition.Y - Size.Y / 2),
-                (int)Size.X,
-                (int)Size.Y,
+                RealPosition.X - Size.X / 2,
+                RealPosition.Y - Size.Y / 2,
+                Size.X,
+                Size.Y,
                 new Color(0, 0, 0, 128),
                 InstructionSource.UI,
                 ZLayer + 0.00004f
